fix: match openui-cloud example to template
